+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.19] - 2026-02-22
+
+- Added support for AWS Polly Long-Form and Generative voices
+- Fix a bug that prevented Neural voices from being used
+
 ## [0.0.18] - 2026-02-22
 
 ### Changed